requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements.txt --locked --no-emit-local --no-all-extras --no-hashes --output-file requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via ewelink-api
aiosignal==1.4.0
    # via aiohttp
attrs==25.4.0
    # via aiohttp
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.11
    # via yarl
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
python-dotenv==1.2.1
    # via ewelink-api
typing-extensions==4.15.0 ; python_full_version < '3.13'
    # via aiosignal
yarl==1.22.0
    # via aiohttp
